backend/requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.20
alembic==1.11.3
psycopg2-binary==2.9.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==1.10.13
pydantic[email]==1.10.13
slowapi==0.1.9
httpx==0.24.1
email-validator==2.0.0.post2
